Refaktorert arkivering typen for å støtte arkivmeldingOppdatering
