+++ dotnet-source/ks.fiks.io.arkivsystem.sample/Handlers/ArkivmeldingHandler.cs
@@ -3,31 +3,32 @@ using System.IO;
 using System.Reflection;
 using System.Xml.Schema;
 using System.Xml.Serialization;
+using KS.Fiks.IO.Arkiv.Client.Models.Arkivering.Arkivmelding;
 using KS.Fiks.IO.Arkiv.Client.Models.Innsyn.Hent.Journalpost;
 using KS.Fiks.IO.Client.Models;
 using Serilog;
 
 namespace ks.fiks.io.arkivsystem.sample.Handlers
 {
-    public class JournalpostHentHandler : BaseHandler
+    public class ArkivmeldingHandler : BaseHandler
     {
         private static readonly ILogger Log = Serilog.Log.ForContext(MethodBase.GetCurrentMethod()?.DeclaringType);
 
-        public static JournalpostHent GetPayload(MottattMeldingArgs mottatt, XmlSchemaSet xmlSchemaSet,
+        public static Arkivmelding GetPayload(MottattMeldingArgs mottatt, XmlSchemaSet xmlSchemaSet,
             out bool xmlValidationErrorOccured, out List<List<string>> validationResult)
         {
             if (mottatt.Melding.HasPayload)
             {
                 var text = GetPayloadAsString(mottatt, xmlSchemaSet, out xmlValidationErrorOccured,
                     out validationResult);
-                Log.Information("Parsing journalpostHent: {Xml}", text);
+                Log.Information("Parsing arkivmelding: {Xml}", text);
                 if (string.IsNullOrEmpty(text))
                 {
-                    Log.Error("Tom journalpostHent? Xml: {Xml}", text);
+                    Log.Error("Tom arkivmelding? Xml: {Xml}", text);
                 }
 
                 using var textReader = (TextReader)new StringReader(text);
-                return(JournalpostHent) new XmlSerializer(typeof(JournalpostHent)).Deserialize(textReader);
+                return(Arkivmelding) new XmlSerializer(typeof(Arkivmelding)).Deserialize(textReader);
             }
 
             xmlValidationErrorOccured = false;
