+++ dotnet-source/ks.fiks.io.arkivsystem.sample/Models/Melding.cs
@@ -1,0 +1,9 @@
+namespace ks.fiks.io.arkivsystem.sample.Models
+{
+    public class Melding
+    {
+        public string FileName { get; set; }
+        public string MeldingsType { get; set; }
+        public object ResultatMelding { get; set; }
+    }
+}
