+++ dotnet-source/ks.fiks.io.fagsystem.arkiv.sample/ks.fiks.io.fagsystem.arkiv.sample.csproj
@@ -26,7 +26,7 @@
 
   <ItemGroup>
     <PackageReference Include="KS.Fiks.ASiC-E" Version="1.0.3" />
-    <PackageReference Include="KS.Fiks.IO.Arkiv.Client" Version="2.0.14" />
+    <PackageReference Include="KS.Fiks.IO.Arkiv.Client" Version="2.0.16-build.20220324103323447" />
     <PackageReference Include="KS.Fiks.IO.Client" Version="1.2.4" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
     <PackageReference Include="Serilog" Version="2.10.0" />
